requirements.txt
Faker==37.12.0
apache-airflow-providers-postgres
apache-airflow-providers-fab
matplotlib

